requirements.txt
requests>=2.28.0
pydantic>=2.0.0
langchain-core>=0.1.0
groq>=0.4.0
google-cloud-translate>=3.8.0
langgraph>=0.0.15
sentence-transformers>=2.2.0
chromadb>=0.4.0
python-dotenv>=1.0.0
pytest>=7.0.0